requirements.txt
azure-functions==1.23.0
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
MarkupSafe==3.0.2
requests==2.32.3
slack_sdk==3.35.0
urllib3==2.4.0
Werkzeug==3.1.3
